app/requirements.txt
Flask>=3.0
Werkzeug>=3.0
gunicorn==20.1.0
python-dotenv